Updated from gitlab MR 833
